+++ demo/index.ts
@@ -16,7 +16,7 @@ async function start() {
     const plugin = new GltfPlugin(map, {
         modelsLoadStrategy: 'waitAll',
         dracoScriptsUrl: 'libs/draco/',
-        ambientLight: { color: 'white', intencity: 2.5 },
+        ambientLight: { color: '#ffffff', intencity: 2.5 },
     });
 
     plugin
